+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - **BC break**: Split subject-line preview SDK out to
   [`phlib/litmus-subjectline-preview`](https://github.com/phlib/litmus-subjectline-preview)
   as it doesn't use the reseller API.
+- **BC break**: Remove `EmailTest::initializeFreeTest()` as one of the clients
+  no longer exists, and there's no documentation to confirm this.
 - **BC break**: Remove PageTest classes as this isn't in the Litmus API.
 - **BC break**: Remove Base* classes. This should not affect implementations,
   which would be using the child `Email*` or `Spam*` classes instead.